+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### Features
 * **terra-form** set empty array as default value for list controls.
+* **terra-checkbox** temporarily added `notifyOnChanges` to be notified on leaf change
+* **terra-checkbox-tree** 
+	- CSS changes
+	- added functionality for disabled leaf
 
 <a name="3.7.0"></a>
 # 3.7.0 (27.03.2019)